# openapi_live.json
{"openapi":"3.1.0","info":{"title":"Sentrix Core Platform","description":"Unified Sentrix Cybersecurity Platform","version":"7.0.0"},"paths":{"/api/v1/threat/events/ingest":{"post":{"tags":["Threat"],"summary":"Ingest Event","operationId":"ingest_event_api_v1_threat_events_ingest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/threat/status":{"get":{"tags":["Threat"],"summary":"Get Threat Status","operationId":"get_threat_status_api_v1_threat_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/predictions/forecast":{"get":{"tags":["Prediction"],"summary":"Get Forecast","operationId":"get_forecast_api_v1_predictions_forecast_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/predictions/forecast/{source_ip}":{"get":{"tags":["Prediction"],"summary":"Get Forecast For Ip","operationId":"get_forecast_for_ip_api_v1_predictions_forecast__source_ip__get","parameters":[{"name":"source_ip","in":"path","required":true,"schema":{"type":"string","title":"Source Ip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/predictions/history":{"get":{"tags":["Prediction"],"summary":"Get History","operationId":"get_history_api_v1_predictions_history_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/investigations/generate/{source_ip}":{"post":{"tags":["Investigation"],"summary":"Generate Investigation","operationId":"generate_investigation_api_v1_investigations_generate__source_ip__post","parameters":[{"name":"source_ip","in":"path","required":true,"schema":{"type":"string","title":"Source Ip"}},{"name":"severity","in":"query","required":false,"schema":{"type":"string","default":"high","title":"Severity"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/investigations/incidents":{"get":{"tags":["Investigation"],"summary":"Get Incidents","operationId":"get_incidents_api_v1_investigations_incidents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/incidents/{incident_id}":{"get":{"tags":["Incident"],"summary":"Get Incident","operationId":"get_incident_api_v1_incidents__incident_id__get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/incidents/{incident_id}/export/json":{"get":{"tags":["Incident"],"summary":"Export Json","operationId":"export_json_api_v1_incidents__incident_id__export_json_get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/incidents/{incident_id}/export/pdf":{"get":{"tags":["Incident"],"summary":"Export Pdf","operationId":"export_pdf_api_v1_incidents__incident_id__export_pdf_get","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/incidents/{incident_id}/assign":{"post":{"tags":["Incident"],"summary":"Assign Incident","operationId":"assign_incident_api_v1_incidents__incident_id__assign_post","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}},{"name":"assignee","in":"query","required":true,"schema":{"type":"string","title":"Assignee"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/incidents/{incident_id}/close":{"post":{"tags":["Incident"],"summary":"Close Incident","operationId":"close_incident_api_v1_incidents__incident_id__close_post","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/incidents/{incident_id}/reopen":{"post":{"tags":["Incident"],"summary":"Reopen Incident","operationId":"reopen_incident_api_v1_incidents__incident_id__reopen_post","parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/":{"get":{"tags":["Rules"],"summary":"List Rules","operationId":"list_rules_api_v1_rules__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Rules"],"summary":"Create Rule","operationId":"create_rule_api_v1_rules__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RulePayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/{rule_id}":{"get":{"tags":["Rules"],"summary":"Get Rule","operationId":"get_rule_api_v1_rules__rule_id__get","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"string","title":"Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Rules"],"summary":"Update Rule","operationId":"update_rule_api_v1_rules__rule_id__put","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"string","title":"Rule Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RulePayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Rules"],"summary":"Delete Rule","operationId":"delete_rule_api_v1_rules__rule_id__delete","parameters":[{"name":"rule_id","in":"path","required":true,"schema":{"type":"string","title":"Rule Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/rules/test":{"post":{"tags":["Rules"],"summary":"Test Rule","operationId":"test_rule_api_v1_rules_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connectors/":{"get":{"tags":["Connectors"],"summary":"List Connectors","operationId":"list_connectors_api_v1_connectors__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/connectors/{connector_name}/enable":{"post":{"tags":["Connectors"],"summary":"Enable Connector","operationId":"enable_connector_api_v1_connectors__connector_name__enable_post","parameters":[{"name":"connector_name","in":"path","required":true,"schema":{"type":"string","title":"Connector Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connectors/{connector_name}/disable":{"post":{"tags":["Connectors"],"summary":"Disable Connector","operationId":"disable_connector_api_v1_connectors__connector_name__disable_post","parameters":[{"name":"connector_name","in":"path","required":true,"schema":{"type":"string","title":"Connector Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connectors/enrich/{ip}":{"post":{"tags":["Connectors"],"summary":"Enrich Ip","operationId":"enrich_ip_api_v1_connectors_enrich__ip__post","parameters":[{"name":"ip","in":"path","required":true,"schema":{"type":"string","title":"Ip"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/health":{"get":{"tags":["Admin"],"summary":"Get Health","operationId":"get_health_api_v1_admin_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/admin/ready":{"get":{"tags":["Admin"],"summary":"Get Ready","operationId":"get_ready_api_v1_admin_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/admin/auth/token":{"post":{"tags":["Admin"],"summary":"Issue Token","description":"Issue a JWT token for admin API key holders.","operationId":"issue_token_api_v1_admin_auth_token_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/suppression/":{"post":{"tags":["Suppression"],"summary":"Create Suppression","operationId":"create_suppression_api_v1_suppression__post","security":[{"APIKeyHeader":[]},{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuppressionPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Suppression Api V1 Suppression  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Suppression"],"summary":"List Suppressions","operationId":"list_suppressions_api_v1_suppression__get","security":[{"APIKeyHeader":[]},{"HTTPBearer":[]}],"parameters":[{"name":"suppression_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Suppression Type"}},{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Active Only"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"},"title":"Response List Suppressions Api V1 Suppression  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/suppression/{suppression_id}":{"delete":{"tags":["Suppression"],"summary":"Delete Suppression","operationId":"delete_suppression_api_v1_suppression__suppression_id__delete","security":[{"APIKeyHeader":[]},{"HTTPBearer":[]}],"parameters":[{"name":"suppression_id","in":"path","required":true,"schema":{"type":"string","title":"Suppression Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Delete Suppression Api V1 Suppression  Suppression Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/suppression/maintenance":{"post":{"tags":["Suppression"],"summary":"Create Maintenance Window","operationId":"create_maintenance_window_api_v1_suppression_maintenance_post","security":[{"APIKeyHeader":[]},{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaintenancePayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Create Maintenance Window Api V1 Suppression Maintenance Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Suppression"],"summary":"List Maintenance Windows","operationId":"list_maintenance_windows_api_v1_suppression_maintenance_get","security":[{"APIKeyHeader":[]},{"HTTPBearer":[]}],"parameters":[{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Active Only"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"},"title":"Response List Maintenance Windows Api V1 Suppression Maintenance Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/soar/audit":{"get":{"tags":["SOAR"],"summary":"Get SOAR Action Audit Log","operationId":"get_audit_log_api_v1_soar_audit_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/soar/execute":{"post":{"tags":["SOAR"],"summary":"Manually trigger a SOAR action (Simulated)","operationId":"trigger_action_api_v1_soar_execute_post","parameters":[{"name":"action_type","in":"query","required":true,"schema":{"type":"string","title":"Action Type"}},{"name":"target","in":"query","required":true,"schema":{"type":"string","title":"Target"}},{"name":"incident_id","in":"query","required":false,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ready":{"get":{"tags":["Observability"],"summary":"Readiness Probe","operationId":"readiness_probe_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/metrics":{"get":{"tags":["Observability"],"summary":"Metrics Probe","operationId":"metrics_probe_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"EventPayload":{"properties":{"event":{"type":"object","title":"Event"}},"type":"object","required":["event"],"title":"EventPayload"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MaintenancePayload":{"properties":{"name":{"type":"string","title":"Name"},"start_time":{"type":"string","title":"Start Time"},"end_time":{"type":"string","title":"End Time"},"scope":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scope","default":"global"},"scope_value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Scope Value"}},"type":"object","required":["name","start_time","end_time"],"title":"MaintenancePayload"},"RulePayload":{"properties":{"rule":{"type":"object","title":"Rule"}},"type":"object","required":["rule"],"title":"RulePayload"},"SuppressionPayload":{"properties":{"suppression_type":{"type":"string","title":"Suppression Type"},"value":{"type":"string","title":"Value"},"rule_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rule Id"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason","default":""},"expires_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expires At"}},"type":"object","required":["suppression_type","value"],"title":"SuppressionPayload"},"TestPayload":{"properties":{"rule":{"type":"object","title":"Rule"},"event":{"type":"object","title":"Event"}},"type":"object","required":["rule","event"],"title":"TestPayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"},"HTTPBearer":{"type":"http","scheme":"bearer"}}}}